Fix for issue 2870: status and dsreplication status do not show replication monitoring The changes consist basically in using the new attributes "approx-older-change-not-synchronized-millis" and "missing-changes" in status and replication status command-lines.  In fact a millisecond version of approx-older-change-not-synchronized-millis has been added in order the remote command-line to be able to provide a localized version of the date (instead of directly presenting the date that was written by the server).
The code has been updated to take into consideration the fact that there is information for several servers; so the replication domain id is used to discriminate the information of the different servers.

+++ opends/src/guitools/org/opends/guitools/statuspanel/BaseDNDescriptor.java
@@ -22,7 +22,7 @@
  * CDDL HEADER END
  *
  *
- *      Portions Copyright 2007 Sun Microsystems, Inc.
+ *      Portions Copyright 2007-2008 Sun Microsystems, Inc.
  */
 
 package org.opends.guitools.statuspanel;
@@ -59,7 +59,7 @@ public class BaseDNDescriptor implements Comparable
   private int nEntries;
   private int missingChanges;
   private DatabaseDescriptor db;
-  private int ageOfOldestMissingChange;
+  private long ageOfOldestMissingChange;
   private Type type;
   private String baseDn;
   private String unescapedDn;
@@ -77,7 +77,7 @@ public class BaseDNDescriptor implements Comparable
    * @param missingChanges the number of missing changes.
    */
   public BaseDNDescriptor(Type type, String baseDn, DatabaseDescriptor db,
-      int nEntries, int ageOfOldestMissingChange, int missingChanges)
+      int nEntries, long ageOfOldestMissingChange, int missingChanges)
   {
     this.baseDn = baseDn;
     this.db = db;
@@ -191,7 +191,7 @@ public class BaseDNDescriptor implements Comparable
    * @return the age of the oldest missing change in seconds in the
    * replication topology for this base DN.
    */
-  public int getAgeOfOldestMissingChange()
+  public long getAgeOfOldestMissingChange()
   {
     return ageOfOldestMissingChange;
   }
